Factor out streaming query builder from query node.
