+++ searchlib/src/vespa/searchlib/query/streaming/CMakeLists.txt
@@ -13,6 +13,7 @@ vespa_add_library(searchlib_query_streaming OBJECT
     onear_query_node.cpp
     phrase_query_node.cpp
     query.cpp
+    query_builder.cpp
     querynode.cpp
     querynoderesultbase.cpp
     queryterm.cpp
